progress: corechip views and models passing
With this commit we now have the post request,
get request methods of the corechip view implemented.
Therefore the backend functionality is working.

Frontend still needs to be properly implemented.
Other minor changes here and there were also added accidentally,
and we keep it as part of this progress commit as they need to be
commited anyways and this is just work in progress.